refactor(advisor): Move the API for plugins to a dedicated module
This allows plugins to only depend on the API without the rest of the
`advisor` module.
