+++ plugins/advisors/api/src/main/kotlin/AdviceProvider.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.advisor
+package org.ossreviewtoolkit.plugins.advisors.api
 
 import org.ossreviewtoolkit.model.AdvisorDetails
 import org.ossreviewtoolkit.model.AdvisorResult
