+++ plugins/advisors/black-duck/build.gradle.kts
@@ -25,8 +25,9 @@ plugins {
 dependencies {
     api(libs.blackduck.common.api)
 
-    api(projects.advisor)
-    api(projects.model)
+    api(projects.plugins.advisors.advisorApi)
+
+    ksp(projects.plugins.advisors.advisorApi)
 
     implementation(libs.blackduck.common)
     implementation(libs.kotlinx.coroutines)
@@ -36,6 +37,4 @@ dependencies {
     funTestImplementation(libs.gson)
 
     testImplementation(libs.gson)
-
-    ksp(projects.advisor)
 }
